src/backend/requirements.txt
fastapi
uvicorn
motor
pymongo
pydantic
python-dotenv
httpx
requests